+++ src/day20.rs
@@ -1,6 +1,6 @@
 use std::fs;
 
-pub fn solve() -> (u32, u32){
+pub fn solve() -> (u64, u64){
     let mut sol1 = 0;
     let mut sol2 = 0;
     panic!("Not implemented!");
